server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"

[dependency-groups]
dev = [
    "ruff>=0.9.3",
]
